EFRAG group printing
Proof-of-concept that EFRAG can be written to an input file using
OBPairData.  This will require that each separate molecule have a really
unique FRAGNAME only for the purpose of writing since I probably do not
want to completely rely on OpenBabel's separator.
Certainly, appending an unique string when reading an input is going to
work, translating from a single XYZ file will need some more thought.